migrations: Use 'datetime' type in MySQL.
Previously DatetimeColumn always used 'timestamp', which is the data type for
date + time in PostgreSQL.

However, in MySQL, it has several "fun" semantics: it will automatically change
to NOW if you don't update it, and it defaults to NOT NULL, unlike every other
data type (that I am currently aware of).

'datetime' seems like MySQL's date + time data type of least surprise, so use
that instead.