domain: Use toJdbcValue with ResultSet instead of setJdbcValue
rs.getObject returns an Integer for some fields where joist expects a Long. Before this commit, this limitation was handled in DomainObjectMapper by using an abstract setJdbcValue method on AliasColumn that takes the ResultSet and a domain instance as parameters and uses the more specific ResultSet methods (e.g. getLong, getDate, etc) as needed. However, this means that the DataTransferObjectMapper cannot map these as expected (e.g. Long for ID) because there is no DomainObject to call setJdbcValue with. Instead, I replaced the abstract setJdbcValue with toJdbcValue that takes in the ResultSet and returns a value of the right type. toJdbcValue can be then be used in both DomainObjectMapper and DataTransferObject with the default implementation of setJdbcValue.