domain: Add sum as aggregate function on AliasColumn
This provides an alternative to SelectCountBuilder and extends the functionality to allow aggregate functions that require columns. It only provides sum, but should be easily extensible to other aggregate functions. It adds an abstract super class of AliasColumn called ColumnExpression and a subclass of ColumnExpression Aggregate. ColumnExpression implements enough of the AliasColumn interface to let it be used in SelectItem and DataTransferObjectMapper without needing the expression to tied to a subclass of DomainObject.

I didn't remove SelectCountBuilder since that would break an existing external API and I'm not sure how Joist handles such in terms of version numbers and the like. It is possible for both implementations to coexist.