migrations: Fix builder 'with' methods to look at sub/super classes.
Previously builders would get a "with(String)" if the String property was unique
within that single entity. However, if super/sub classes also had single String
properties, then they would also get "with(String)" methods.

Same this for code values, like builder.isCodeValue().

This pulls the calculation of "is this type unique to the root entity and all of
it's sub entities" into the Entity class.