migrations: Add covariant overrides for fluent setters.
Previously, if a base entity builder added a fluent setter, the child
entity builder would get that method for free.

However, the return type was still the base class's builder type, which meant if
a caller was trying to chain together calls against the child builder type, the
return type wouldn't suddenly jump to the parent builder type.

So, to avoid this, we just brute-over generate overrides for each fluent
setter in the sub entity so that we can have it return the right type.