migrations: Consider uniques within the branch, not the whole tree.
Previously, given ParentA, ChildA extends ParentA, and ChildB extends ParentA,
if ChildA and ChildB both had a string type, it would not be considered unique,
because we went up to ParentA and asked it for unique types within the whole
inheritance tree.

However, since ChildA and ChildB are in different branches (not sub or super
classes of each other), their types should not conflict.

So, now the unique calc only looks at the current class's base classes, current
class, and all subclasses.