util: Remove Interpolate.string calls from Jdbc utility class.
The idea was that users could be trusted to know when it was/was not safe to use
Jdbc's various "Object..." args parameters.

That is not the case.

The assumption was originally formed when the primary user of joist-util was
only joist-migrations, which has 100% control of the input, and needed to
interpolate values into locations (e.g. table name for creating/dropping tables)
not allowed by JDBC.

This made it seem like a good idea for raw string interpolation to happen
inside of Jdbc.

But after the users of Jdbc grew to be more than just joist-migrations,
it proved, of course, to be a very bad idea.