domain: Fix NPE when querying only one column with Select.list(Class).
DataTransferObjectMapper was trying to be cute and infer whether
the user was loading a DTO or a value by looking at the type
of the loaded value (and only when querying a single column).

However, if that value is null, not only does our value.getClass
NPE, but we fundamentally can't tell whether the user wanted to
load our DTO (so make a new DTO instance with no field set) or
wanted the value (just add null to the list).

So, we introduce a new .listValues(Class) that is specifically
meant for querying only one field.