migrations: Always add XxxAlias.on/leftOn.
This is a more natural way to do it (vs. the previous ugly checks
of 'have we added this method yet?'), plus also handles joining
into an entity that has had it's parent/child collection marked as
skipped due to likely being too large.

Which actually makes the liklihood that a user will want to join
into the alias higher than it would be otherwise.